test(suite-native): stub looping videos in E2E builds
TrezorAnimation, BluetoothPairingAnimation and Video all play looping
H.264 assets, the largest of them 2000x2666 at 30 fps. The Android
emulator decodes those in its own host process, which starves the
emulated CPU and is a known source of emulator crashes on CI runners.
E2E builds now only reserve the layout space instead of playing anything.
