+++ suite-native/module-authorize-device/src/components/connect/BluetoothPairingAnimation.styles.ts
@@ -1,0 +1,6 @@
+import { prepareNativeStyle } from '@trezor/styles-native';
+
+export const animationStyle = prepareNativeStyle(({ borders }) => ({
+    aspectRatio: 1,
+    borderRadius: borders.radii.r16,
+}));
