+++ suite-native/module-authorize-device/src/components/connect/BluetoothPairingAnimation.tsx
@@ -3,18 +3,15 @@ import React from 'react';
 import { VideoView, useVideoPlayer } from 'expo-video';
 
 import { useActiveColorScheme } from '@suite-native/theme';
-import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
+import { useNativeStyles } from '@trezor/styles-native';
 import { type ThemeColorVariant } from '@trezor/theme';
+
+import { animationStyle } from './BluetoothPairingAnimation.styles';
 
 const bluetoothPairingAnimations = {
     standard: require('../../assets/bluetooth-pairing-standard.mp4'),
     dark: require('../../assets/bluetooth-pairing-dark.mp4'),
 } as const satisfies Record<ThemeColorVariant, string>;
-
-const animationStyle = prepareNativeStyle(({ borders }) => ({
-    aspectRatio: 1,
-    borderRadius: borders.radii.r16,
-}));
 
 export const BluetoothPairingAnimation = () => {
     const bluetoothPairingAnimation = bluetoothPairingAnimations[useActiveColorScheme()];
